+++ tests/testthat/test-class_fst.R
@@ -19,24 +19,6 @@ tar_test("fst format", {
   expect_silent(target_validate(x))
 })
 
-tar_test("trust_object_timestamps = TRUE", {
-  skip_if_not_installed("fst")
-  old <- tar_option_get("trust_object_timestamps")
-  on.exit(tar_option_set(trust_object_timestamps = old))
-  tar_option_set(trust_object_timestamps = TRUE)
-  x <- target_init(name = "abc", expr = quote(a), format = "fst")
-  expect_true(x$store$file$trust_timestamps)
-})
-
-tar_test("trust_object_timestamps = FALSE", {
-  skip_if_not_installed("fst")
-  old <- tar_option_get("trust_object_timestamps")
-  on.exit(tar_option_set(trust_object_timestamps = old))
-  tar_option_set(trust_object_timestamps = FALSE)
-  x <- target_init(name = "abc", expr = quote(a), format = "fst")
-  expect_false(x$store$file$trust_timestamps)
-})
-
 tar_test("non-data-frame throws an error", {
   skip_if_not_installed("fst")
   tar_script(